Add missing unit tests ProcessBillingBatchService
https://eaflood.atlassian.net/browse/WATER-3803

When we create the `ProcessBillingBatchService` we did it as a spike, as we were unsure of what the final service would look like. It's been through some changes since then, especially once we added support for crediting the previous billing batch in.

But now we think we have a stable design that we will iterate rather than trash and re-write in future changes. So, as promised, it is time to start adding the missing unit tests.