Add a missing unit test for process billing
Whilst thinking of the scenarios to test for ProcessBillingBatchService we though of one not currently tested for the ProcessBillingTransactionsService.

What happens if we calculate no billable transactions but the previous billing batch does have transactions for the matching invoice and licence? Our current understanding is that we would credit everything on the previous bill run so our test ensures we get that result.