Fetch all debit trans. from previous bill runs
https://eaflood.atlassian.net/browse/WATER-3950

During the supplementary billing process as we iterate through the charge versions we do so in invoice account then licence order. This is so we can use 'iterating to the next one' as a trigger to know when to persist the current billing invoice and billing licence records we're dealing with.

Because we have the invoice account and licence ID's during this process it made it simple to look at the last bill run and extract any matching debit billing transactions that might now need crediting.

But during test a scenario came up that we're not managing. A charge version can be created that supercedes the previous one. It supercedes the previous, because the the start date of the new charge version is the same as the previous i.e. the previous charge version should be completely ignored.

This most often happens because the billing account was wrong. So, the new charge version will be linked to a different billing account. The problem is when the previous charge version has already been billed. We need to include a credit in the new bill run we're generating for it. But our 'prudent' inclusion of the invoice account ID in our query means we're overlooking the previous debit to the old billing invoice account.

The end result is 2 invoices to 2 billing accounts on different bill runs, for the same licence and abstraction period.

This change resolves the problem, though goodness knows how!