Remove FetchLicencesService (#178)
https://eaflood.atlassian.net/browse/WATER-3948

To resolve the issue of dealing with PRESROC and SROC licences flagged for supplementary billing, and how we remove those flags when bill runs get 'sent' we've [made a change to the `FetchChargeVersionsService`](https://github.com/DEFRA/water-abstraction-system/pull/177) to look at a [new field we've added](https://github.com/DEFRA/water-abstraction-service/pull/2077) to the `water.licences` table.

This broke some tests, a number of which were in the `FetchLicencesService`. We were about to embark on fixing them when we remembered that with the direction our work on supplementary billing took `FetchLicencesService` had become redundant.

So, rather than spend time fixing them we're just getting on and removing the service.