We are incorrectly crediting all the previous charge versions (#182)
https://eaflood.atlassian.net/browse/WATER-3951

When running a supplementary bill run, the service that fetches the previous billing transactions is not just returning the transactions from the previous bill run. It is returning the transactions from all previous bill runs that match the financial year ending, licence & invoice account.

This fix ensures that only data from the previous bill run is returned.