Hack for supplementary billing being late
SROC supplementary billing was always intended to be shipped for use in financial year 2022-23. We know at some point it will need to go back and look at previous years, as the PRE-SROC legacy version already does. But when we started this functionality was going to be added later.

So, we wrote the current functionality to just look at the current financial year. That's a great idea when it's October 2022 and you aim to be done by Christmas! Fast forward to now, we're still plugging away and finding scenarios we've not covered or not handled (because we didn't know BTW). Which means we've fallen into the financial year 2023-24!

😱😱😱😱😱

As a hack, to allow us to ship supplementary billing and get the 2022-23 bills out and allow us to then move to multi-year billing we're hard coding the `BillingPeriodService` to always return 2022-23 (the first year of SROC's existence).