Add new FetchReplacedChargeVersionsService
Modelled on `FetchChargeVersionsService` this returns those charge versions that display as REPLACED in the UI (superseded) in the DB.

We don't return the same amount of data and linked records. Unlike the `FetchChargeVersionsService`, the results we return won't be used as the basis for calculating the transaction lines. Instead, they'll be used to identify existing billing transaction lines.