Update FetchReplacedChargeVersionsService
In testing we hit on a scenario where the `FetchReplacedChargeVersionsService` might pull through charge versions with invoice account ID's that have already been processed as part of the main `ProcessBillingBatchService`.

When this happened an error was thrown because it tried to insert into `billing_invoices` a record with the same invoice account ID and billing batch ID as already exists.

If a replaced charge version is linked to an invoice account that is also linked to an approved invoice account it will already have been processed.