Working implementation
Alongside the other services we have added, this change provides a working solution to the replaced charge versions problem.

I have tested multiple replacements of a charge version where the billing account has been changed, within a single licence, and for partial parts of the year.

It seems to be working!

This started as an exercise in getting the sudo-code PR into something that more closely resembled what is needed. But it's ended up with a working solution.

There is probably some more refactoring we can do to improve things. We could certainly in more documentation and missing unit tests.

But hopefully what is here will provide a base on which to make those changes.