Standardise how we test the length of an array (#190)
In our unit tests, we had been checking the length of an array using the method `expect(result.length).to.equal(1)`. However, our preferred way to test the length of an array is `expect(result).to.have.length(1)`.

This PR updates all unit tests to use the `expect(result).to.have.length(1)` method to test the length of an array.