Fix 'draft' tests
One of our failing tests highlighted we had been a bit to eager when deleting the status WHERE condition. Though we've not seen much evidence of its use, charge versions do have a status of 'draft'.

So, just in case we add that WHERE condition back in and update the tests appropriately.