Make test helpers consistent (#198)
To support our unit testing we have a convention of creating ‘helpers’ that create the data we need to test our code.

Often data is related, for example, a `ChargeVersion` is linked to a `Licence` record. When we first built the helpers we copied the model used by [sroc-charging-module-api](https://github.com/DEFRA/sroc-charging-module-api), which will create linked data records automatically.

But as time went on, we realised the complexity of the WRLS data schema was making this pattern unworkable so we stopped. But our initial data helpers still follow that pattern. This has caused us to waste time or get confused at times when working on SROC supplementary billing.

So, before we start the next phase of work, we need to update the original helpers to stop creating linked records and update existing tests accordingly.