Extract table results to a file (#193)
https://eaflood.atlassian.net/browse/WATER-3964

This pull request is just a small part of a larger project that involves exporting all our database schemas, converting them into CSV files, and uploading them to our Amazon S3 bucket. 
Its primary focus is exporting the billing charge categories table to a file. Converting it to a CSV standard first. 
To expedite the export process and see the output sooner, we are using a vertical slicing approach, rather than a horizontal one, which means exporting a single table at a time from each schema.