https://eaflood.atlassian.net/browse/WATER-3965
This pull request is just a small part of a larger project that involves exporting all our database schemas, converting them into CSV files, and uploading them to our Amazon S3 bucket.
This PR's primary focus is compressing the billing charge categories table that we have already converted and exported to a CSV file.
To expedite the export process and see the output sooner, we are using a vertical slicing approach, rather than a horizontal one, which means exporting a single table at a time from each schema.