Create new generator functions in test
We need to add new tests to cover checking when the agreements and charges differ between previous and calculated transactions.

It's going to need more transaction data to be passed through to the service, and continuing to use fixed data is going to make the test file massive.

So, we make a change to move the test data logic to be generator methods instead.