Tidy supplementary billing code - second pass
We've been working quickly to get supplementary billing completed on time. This means we haven't always been able to spend as much time as we'd like on some of the finer details -- documentation, consistency of standards, naming conventions, etc. This PR is a second pass at spotting these things and amending them.