Simplify process of finding latest end date
In `DetermineChargePeriodService` we were creating a variable for each timestamp, defaulting to `financialYearEndDate` if the timestamp is `null`, then using `Math.min()` to find the earliest of these -- the defaulting is required as a `null` value would be counted as `0` by `Math.min()`. We can streamline this and cut the number of lines required by simply putting the original values into an array then filtering out the `null` values, then passing our array values to `Math.min()`.