Tidy supplementary billing code - second pass (#208)
We've been working quickly to get supplementary billing completed on time. This means we haven't always been able to spend as much time as we'd like on some of the finer details -- documentation, consistency of standards, naming conventions, etc. [Following our first pass](https://github.com/DEFRA/water-abstraction-system/pull/197), this PR is a second pass at spotting these things and amending them.

Note that while refactoring `ProcessBillingBatchService` we identified a couple of ways we can restructure the process. This will be handled in a separate PR; for now, we have simply refactored code into helper functions which more accurately describe the flow of the process.