Fix incorrect cleanse of previous transactions
> https://eaflood.atlassian.net/browse/WATER-4005

We have found another issue in UAT for the SROC supplementary billing. When fetching previous billing transactions we compare the debits against the credits and 'cleanse' any that match. This is how we determine if there is anything previous to be credited in the new bill run, because they will be the debits that didn't match to anything.

But we have found in testing there is a scenario where there is a debit to be credited but we're letting the _wrong_ debit billing transactions through. This is because we are only comparing charge type and number of billable days. We need to be looking at a number of other properties.

To explain the issue, we've put together this example.

**Before the fix**

So, imagine we have licence `01/123/456` and linked to it are 1 new SROC charge version with no additional agreements.

> For the purposes of this the abstraction period is all year and the charge reference is the same for each charge version. Also, the charge version ID below ascends in the order they are added. But we list them in the tables in the order the UI will display them.

***Bill run 1***

|ID   |Add. agreements?|Start|End  |Status  |
|-----|----------------|-----|-----|--------|
|CHG01|-               |01/04|-    |APPROVED|

The annual bill run was generated resulting in

|ID   |Charge Version|Add. agreements?|Billable days|Debit or Credit?|Amount|
|-----|--------------|----------------|-------------|----------------|------|
|TRN01|CHG01         |-               |365          |Debit           |1000  |

***Bill run 2***

Then, a new charge version with an additional agreement is added which replaces `CHG01`.

|ID   |Add. agreements?|Start|End  |Status  |
|-----|----------------|-----|-----|--------|
|CHG02|Winter discount |01/04|-    |APPROVED|
|CHG01|-               |01/04|-    |REPLACED|

The supplementary bill run is generated resulting in

|ID   |Charge Version|Add. agreements?|Billable days|Debit or Credit?|Amount|
|-----|--------------|----------------|-------------|----------------|------|
|TRN02|CHG01         |-               |365          |Credit          |1000  |
|TRN03|CHG02         |Winter discount |365          |Debit           |500   |

***Bill run 3***

Then, another charge version is added which replaces `CHG02`. The reason doesn't matter so in our example we just changed the description.

|ID   |Add. agreements?|Start|End  |Status  |
|-----|----------------|-----|-----|--------|
|CHG03|Winter discount |01/04|-    |APPROVED|
|CHG02|Winter discount |01/04|-    |REPLACED|
|CHG01|-               |01/04|-    |REPLACED|

The supplementary bill run is generated. Depending on the order that `FetchPreviousBillingTransactionsService` returns the transactions in we could get one of 2 bill runs (first correct, second incorrect).

|ID   |Charge Version|Add. agreements?|Billable days|Debit or Credit?|Amount|
|-----|--------------|----------------|-------------|----------------|------|
|TRN04|CHG02         |Winter discount |365          |Credit          |500   |
|TRN05|CHG03         |Winter discount |365          |Debit           |500   |

or

|ID   |Charge Version|Add. agreements?|Billable days|Debit or Credit?|Amount|
|-----|--------------|----------------|-------------|----------------|------|
|TRN04|CHG01         |-               |365          |Credit          |1000  |
|TRN05|CHG03         |Winter discount |365          |Debit           |500   |

If the service received the transactions in this order it would result in the second bill run.

|ID   |Charge Version|Add. agreements?|Billable days|Debit or Credit?|Amount|
|-----|--------------|----------------|-------------|----------------|------|
|TRN03|CHG02         |Winter discount |365          |Debit           |500   |
|TRN01|CHG01         |-               |365          |Debit           |1000  |
|TRN02|CHG01         |-               |365          |Credit          |1000  |

This is because `FetchPreviousBillingTransactionsService` splits the results into debits and credits, then compares the 2. Prior to this change we only compared billable days (and charge type but that is irrelevant to the issue). So, our £500 winter discount based debit (`TRN03`) would be cancelled by the £1000 no agreement based credit (`TRN02`).

This would leave the our supplementary billing engine to think `TRN01`, the £1000 debit needs to be credited.

**The fix**

By comparing other properties, including agreements as well as the billable days the correct credit will be cancelled out by the correct debit.