Fetch invoice account numbers instead of invoice accounts
Since we only really need to get the invoice account numbers from the db, we rework `FetchInvoiceAccountsService` to only return the numbers along with the ids for later indexing.

To avoid confusion we amend the service to be `FetchInvoiceAccountNumbers` which makes it very clear that what it returns isn't an array of `InvoiceAccountModel` instances but simply an array of objects which each contain the account number and id.