Refactor `ProcessBillingBatchService` - first pass (#214)
https://eaflood.atlassian.net/browse/WATER-3996

While reviewing `ProcessBillingBatchService` we realised there is scope to make it more performant. This is a first pass which tackles the issue of us fetching invoice accounts one at a time within the main loop. We refactor so that we now fetch all the invoice accounts we will need in a single query prior to entering the main loop and hold them in a keyed object:
```js
{
  '610ca07a-3bd3-4ea2-8cb1-82f6977cdc71': { invoiceAccountId: '610ca07a-3bd3-4ea2-8cb1-82f6977cdc71', ... },
  '03cec07f-9ad3-4c9a-bad8-1958cec0d61f': { invoiceAccountId: '03cec07f-9ad3-4c9a-bad8-1958cec0d61f', ... }
}
```
This means that we can easily retrieve a given invoice account once we're in the main loop:
```js
const invoiceAccountId = invoiceAccounts['610ca07a-3bd3-4ea2-8cb1-82f6977cdc71']
```