Uploading file to AWS S3 bucket (#211)
https://eaflood.atlassian.net/browse/WATER-3966

This pull request is just a small part of a larger project that involves exporting all our database schemas, converting them into CSV files, and uploading them to our Amazon S3 bucket. This PR's primary focus is connecting to the amazon s3 bucket and uploading the file. To expedite the export process and see the output sooner, we are using a vertical slicing approach, rather than a horizontal one, which means exporting a single table at a time from each schema.