Add prefix to 2PT billing transaction descriptions
https://eaflood.atlassian.net/browse/WATER-4006

It has been highlighted in testing that users expect to see a different description for any billing transactions linked to a charge version with a two-part tariff agreement in place.

For example, a standard description would be

> Water abstraction charge: Spray Irrigation at Baygrove Upcot Oldent

If the charge version has a two-part tariff agreement it should be

> Two-part tariff basic water abstraction charge: Spray Irrigation at Baygrove Upcot Oldent

This change makes that happen.