Refactor to remove unnecessary iteration and generation
In our `_preGenerateBillingInvoiceLicences` function we were iterating over all charge versions to generate a billing invoice licence for each one, then iterating over the generated billing invoice licences to produce the final keyed object. We realised that we could combine the two steps by iterating over the charge versions and generating the billing invoices in the same step. This means we can also check whether the billing invoice and licence in question already has a billing invoice licence generated, which we weren't doing with the first approach.