`ProcessBillingBatchService` refactor - pre-generate invoice licences (#218)
https://eaflood.atlassian.net/browse/WATER-3996

In order to fix a bug introduced by our refactoring, we moved to pre-generating all required billing invoices prior to entering the main charge version processing loop. Having done this, we think it would be a good idea to also pre-generate the required invoice licences.

This PR also introduces further optimisations to the pre-generation of the billing invoices, reducing the number of times we need to iterate in order to produce the object containing the generated billing invoices.