First pass at moving db writes outside of loop
Instead of having a `currentBillingData` object which we update as we iterate over the charge versions, we have a `billingData` object which is a keyed equivalent where the key is the billing invoice licence id and the value is effectively the `currentBillingData`. We update this as we iterate over the charge versions, then iterate over it after the loop to perform the db writing. Although this doesn't yet reduce the number of db writes, it does move them outside of the loop. As a bonus, we can simply update the desired billing invoice licence's data on each iteration and we no longer need to check for a change in id