Implement fix
If we pass `undefined` or `null` to `Number()` it will return `NaN` which is interpreted as `false`. So, the result is the same as previous and the default value is returned.

If it has a value though, we'll now have it set as a number ready for comparison with any previous billing transactions, which save them as numbers.