Update charge element helper to match real life
Having spotted this issue we update the charge element helper to default the adjustments to match how they are actually stored in the DB.

This also means we by default test our changes. `aggregate` when converted during testing of `GenerateBillingTransactionsService` should match as a number. `s126` and `charge` left as is will confirm we still use the default when no value exists.