Calculate `isEmpty` and rename to `billingBatchIsEmpty`
When we loop over the billing data, we previously set the `isEmpty` flag to `false` before we add transactions, billing invoices and billing invoice licences to our `dataToPersist` object. Instead of setting the flag we can instead calculate it based on whether `dataToPersist.billingInvoiceLicences` is populated. We also rename it to `billingBatchIsEmpty` to make its intent a little clearer