Check for presence of billing data
Instead of blindly overwriting `billingData[billingInvoiceLicenceId]` with either the data we provided or some empty initial data, we check first to see if it exists and only write our defaults if it doesn't. This makes it clearer what we're doing (ie. "only set the initial data if something doesn't already exist")