Stop sending Airbrake notifications in tests (#229)
Spotted when needing to go into [Errbit](https://errbit.com/) for another reason we are sending _real_ notifications when running our unit tests.

This change ensures it's stubbed out where this is happening to avoid real notifications being sent. Also with this change, we remove the noise that appeared in our test output when working with errors.