Fix issue with testing ProcessBillingBatchService
We really want to test what the ProcessBillingBatchService service _was_ doing in it's `_process()` method. But the way we have designed this is to kick off the processing whilst still returning a response to the client.

This made it extremely painful to test. Then inspiration hit; if we move what `_process()` is doing to its own service we can then test that as normal!

Key thing is that new service really needed to keep the name `ProcessBillingBatchService`! So, we introduce a new service that will be called from the controller, will do all the work of determining the billing periods and creating the initial billing batch record. It will also kick of our ProcessBillingBatchService whilst sending a response back to the controller.