Generalising fetch billing charge categories service (#232)
https://eaflood.atlassian.net/browse/WATER-4015

This pull request is just a small part of a larger project that involves exporting all our database schemas, converting them into CSV files, and uploading them to our Amazon S3 bucket. This PR's primary focus is to generalize the current service we have so that it can take any table as an argument. To expedite the export process and see the output sooner, we are using a vertical slicing approach, rather than a horizontal one, which means exporting a single table at a time from each schema.