Refactor, unit tests & add NewBillingBatchService (#236)
https://eaflood.atlassian.net/browse/WATER-3977
https://eaflood.atlassian.net/browse/WATER-3984

This is a follow on from [Refactor start of processing a supp. billing batch](https://github.com/DEFRA/water-abstraction-system/pull/233). In there we needed to refactor the `ProcessBillingBatchService` to not mix processing a given billing period and the overall billing batch. So, we created `ProcessBillingPeriodService` and split the functionality between the 2.

But we parked writing any unit tests for our new `ProcessBillingBatchService` till this PR, else the change would have been massive. Only we hit an issue when we started writing them.

`ProcessBillingBatchService` in its new form was determining the billing periods, initiating the billing batch record, and generating a response back to the client all whilst kicking off the process for each billing period and handling any errors thrown. It was also responsible for finalising the bill run.

A bit much! The final issue was that background processing was not `awaited`. This made it a nightmare to try and test all this behaviour.

We realised we needed to get the 'processing' bit in its own service if only to make it testable. So, this change adds a new `NewBillingBatchService` (yes - we really need to sit down and firm up our naming conventions!)

_This_ is now what is called when kicking off a new billing batch. It deals with determining the billing periods, initiating the billing batch, and generating the response for the client. Essentially, it manages everything to do with a new bill run. It also kicks off `ProcessBillingBatchService` in the background, which is now solely focused on processing a billing batch; generating the data, handling errors, updating the billing batch status, and finalising it.

With this in place, it meant we could also write some unit tests!